+++ doc/source/CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 **Fixes**:
 
 - msm: Chapman Kolmogorov validator ensures there are no side effects on the tested model. #1255
+- datasets: Fix default values for kT to ensure integrator produces sane values. #1272, #1275
 
 
 2.5.1 (02-17-2018)